// packages/shared/src/types.ts
export type CharacterId =
  | 'gandalf'
  | 'legolas'
  | 'saruman'
  | 'aragorn'
  | 'gimli'
  | 'sam'
  | 'frodo'
  | 'boromir'
  | 'gollum';

export type BookId =
  | 'the-hobbit'
  | 'the-fellowship-of-the-ring'
  | 'the-two-towers'
  | 'the-return-of-the-king'
  | 'the-silmarillion'
  | 'unfinished-tales';

export type Locale = 'en' | 'tr' | 'pt';